+++ changelog.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2.6] - 2021-06-03
+### Updated
+
+- Add support for the `ical-tec` feature plugin.
+
+## [1.2.5] - 2021-05-26
+### Updated
+
+- Update `node` and `gulp` related commands.
+
 ## [1.2.4] - 2020-12-28
 ### Updated